Add a test to verify ViewResolver GetView
