+++ src/Radical.Windows.Tests/Presentation/Regions/Specialized/TacControlRegionTests.cs
@@ -2,6 +2,7 @@
 using Radical.Windows.ComponentModel;
 using Radical.Windows.Presentation;
 using Radical.Windows.Regions;
+using Radical.Windows.Tests;
 using System;
 using System.Threading;
 using System.Windows;
@@ -106,50 +107,6 @@ namespace Test.Radical.Windows.Presentation.Regions.Specialized
                 Assert.IsTrue(vm1.Invoked);
                 Assert.IsTrue(vm2.Invoked);
             });
-        }
-    }
-
-    class TestRunner
-    {
-        readonly Action test;
-        readonly ApartmentState state;
-        Exception ex;
-
-        private TestRunner(ApartmentState state, Action test)
-        {
-            this.state = state;
-            this.test = test;
-        }
-
-        public static void Execute(ApartmentState state, Action test)
-        {
-            var runner = new TestRunner(state, test);
-            runner.Execute();
-        }
-
-        private void Execute()
-        {
-            var worker = new Thread(() =>
-            {
-                try
-                {
-                    this.test();
-                }
-                catch (Exception e)
-                {
-                    Console.WriteLine(e);
-                    this.ex = e;
-                }
-            });
-
-            worker.SetApartmentState(this.state);
-            worker.Start();
-            worker.Join();
-
-            if (this.ex != null)
-            {
-                throw this.ex;
-            }
         }
     }
 }
