+++ src/Radical.Windows.Tests/Services/Presentation/AnEmptyView.xaml
@@ -1,0 +1,11 @@
+﻿<Window x:Class="Radical.Windows.Tests.Services.Presentation.AnEmptyView"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        mc:Ignorable="d"
+        Title="AnEmptyView" Height="450" Width="800">
+    <Grid>
+        
+    </Grid>
+</Window>
