+++ src/Radical.Windows.Tests/Services/Presentation/AnEmptyView.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows;
+
+namespace Radical.Windows.Tests.Services.Presentation
+{
+    /// <summary>
+    /// Interaction logic for AnEmptyView.xaml
+    /// </summary>
+    public partial class AnEmptyView : Window
+    {
+        public AnEmptyView()
+        {
+            InitializeComponent();
+        }
+    }
+}
